Replace polling with event touch class, add chest button events
