+++ edante/sensing/CMakeLists.txt
@@ -95,10 +95,6 @@ add_executable(sensing
  src/fsr.cpp
  )
 
-add_executable(sensing_main
- src/sensing_main.cpp
- src/sensing_test.cpp)
-
 ## Add cmake target dependencies of the executable/library
 ## as an example, message headers may need to be generated before nodes
 add_dependencies(sensing sensing_generate_messages_cpp)
@@ -106,14 +102,6 @@ add_dependencies(sensing sensing_generate_messages_cpp)
 
 ## Specify libraries to link a library or executable target against
 target_link_libraries(sensing
-  ${catkin_LIBRARIES}
-  alcommon
-  alerror
-  alproxies
-  alvalue
-)
-
-target_link_libraries(sensing_main
   ${catkin_LIBRARIES}
   alcommon
   alerror
