+++ edante/sensing/src/sensing.cpp
@@ -2,37 +2,34 @@
 * @Copyright: Copyright[2015]<Alejandro Bordallo>
 * @Date:      2015-05-11
 * @Email:     alex.bordallo@ed.ac.uk
-* @Desc:      ROS Wrapper for NaoQI sensor events
+* @Desc:      Add file description...
 */
-
-#include <ros/ros.h>
-#include "definitions.h"
+#include <signal.h>
+#include <boost/shared_ptr.hpp>
+#include <alcommon/albroker.h>
+#include <alcommon/almodule.h>
+#include <alcommon/albrokermanager.h>
+#include <alcommon/altoolsmain.h>
 
 #include "touch.h"
-#include "power.h"
-#include "sonar.h"
-#include "fsr.h"
 
-#include <alproxies/almemoryproxy.h>
+extern "C"
+{
+  int _createModule(boost::shared_ptr<AL::ALBroker> pBroker) {
+    AL::ALBrokerManager::setInstance(pBroker->fBrokerManager.lock());
+    AL::ALBrokerManager::getInstance()->addBroker(pBroker);
+    AL::ALModule::createModule<Touch>( pBroker, "Touch");
+    return 0;
+  }
+  int _closeModule() {
+    return 0;
+  }
+}
 
 int main(int argc, char *argv[]) {
-  ros::init(argc, argv, "sensing");
-  ros::NodeHandle nh("sensing");
-  AL::ALMemoryProxy memProxy("127.0.0.1", 9559);
-  Touch TouchTest(&nh, &memProxy);
-  // Power PowerTest(&nh, &memProxy);
-  // Sonar SonarTest(&nh, &memProxy);
-  // Fsr FsrTest(&nh, &memProxy);
-
-  ros::Rate r(10);
-
-  while (ros::ok()) {
-    TouchTest.spinTopics();
-    // PowerTest.spinTopics();
-    // SonarTest.spinTopics();
-    // FsrTest.spinTopics();
-    ros::spinOnce();
-    r.sleep();
-  }
-
+  // pointer to createModule
+  TMainType sig;
+  sig = &_createModule;
+  // call main
+  ALTools::mainFunction("Touch", argc, argv, sig);
 }
