+++ edante/sensing/src/touch.cpp
@@ -2,119 +2,265 @@
 * @Copyright: Copyright[2015]<Alejandro Bordallo>
 * @Date:      2015-05-12
 * @Email:     alex.bordallo@ed.ac.uk
-* @Desc:      ROS wrapper for Nao's touch sensors
+* @Desc:      Testing ALModule and ROS topic integration
 */
 
 #include "touch.h"
 
-Touch::Touch(ros::NodeHandle* nh, AL::ALMemoryProxy* memProxy) {
-  nh_ = nh;
-  memProxy_ = memProxy;
-  INFO("Setting up Touch publishers" << std::endl);
+#include <alvalue/alvalue.h>
+#include <alcommon/alproxy.h>
+#include <alcommon/albroker.h>
+#include <althread/alcriticalsection.h>
+
+#include <qi/log.hpp>
+
+Touch::Touch(
+  boost::shared_ptr<AL::ALBroker> broker,
+  const std::string& name): AL::ALModule(broker, name),
+  fCallbackMutex(AL::ALMutex::createALMutex()) {
+  qi::log::setVerbosity(qi::log::silent);
+  // setModuleDescription("Sensing Test module.");
+
+  functionName("rightBumperPressed", getName(), "Right bumper is pressed.");
+  BIND_METHOD(Touch::rightBumperPressed)
+  functionName("leftBumperPressed", getName(), "Left bumper is pressed.");
+  BIND_METHOD(Touch::leftBumperPressed)
+  functionName("frontHeadPressed", getName(), "Front head is pressed.");
+  BIND_METHOD(Touch::frontHeadPressed)
+  functionName("middleHeadPressed", getName(), "Middle head is pressed.");
+  BIND_METHOD(Touch::middleHeadPressed)
+  functionName("rearHeadPressed", getName(), "Rear head is pressed.");
+  BIND_METHOD(Touch::rearHeadPressed)
+  functionName("handRightBackPressed", getName(),
+               "Hand right back is pressed.");
+  BIND_METHOD(Touch::handRightBackPressed)
+  functionName("handRightLeftPressed", getName(),
+               "Hand right left is pressed.");
+  BIND_METHOD(Touch::handRightLeftPressed)
+  functionName("handRightRightPressed", getName(),
+               "Hand right right is pressed.");
+  BIND_METHOD(Touch::handRightRightPressed)
+  functionName("handLeftBackPressed", getName(),
+               "Hand left back is pressed.");
+  BIND_METHOD(Touch::handLeftBackPressed)
+  functionName("handLeftLeftPressed", getName(),
+               "Hand left left is pressed.");
+  BIND_METHOD(Touch::handLeftLeftPressed)
+  functionName("handLeftRightPressed", getName(),
+               "Hand left right is pressed.");
+  BIND_METHOD(Touch::handLeftRightPressed)
+  functionName("singleChestClick", getName(),
+               "Single chest button click event.");
+  BIND_METHOD(Touch::singleChestClick)
+  functionName("doubleChestClick", getName(),
+               "Double chest button click event.");
+  BIND_METHOD(Touch::doubleChestClick)
+  functionName("tripleChestClick", getName(),
+               "Triple chest button click event.");
+  BIND_METHOD(Touch::tripleChestClick)
+  int argc = 0;
+  char *argv[] = {NULL};
+  ros::init(argc, argv, "touch");
+  nh_ = new ros::NodeHandle("touch");
   bumpers_pub_ = nh_->advertise<sensing::bumpers>("bumpers", 10);
-  chest_pub_ = nh_->advertise<std_msgs::Bool>("chest", 10);
+  chest_pub_ = nh_->advertise<std_msgs::String>("chest", 10);
   head_pub_ = nh_->advertise<sensing::head>("head", 10);
   right_hand_pub_ = nh_->advertise<sensing::hand>("rightHand", 10);
   left_hand_pub_ = nh_->advertise<sensing::hand>("leftHand", 10);
 }
 
 Touch::~Touch() {
+  fMemoryProxy.unsubscribeToEvent("RightBumperPressed", "Touch");
+  fMemoryProxy.unsubscribeToEvent("LeftBumperPressed", "Touch");
+  fMemoryProxy.unsubscribeToEvent("FrontTactilTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("MiddleTactilTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("RearTactilTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandRightBackTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandRightLeftTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandRightRightTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandLeftBackTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandLeftLeftTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("HandLeftRightTouched", "Touch");
+  fMemoryProxy.unsubscribeToEvent("ALSentinel/SimpleClickOccured", "Touch");
+  fMemoryProxy.unsubscribeToEvent("ALSentinel/DoubleClickOccured", "Touch");
+  fMemoryProxy.unsubscribeToEvent("ALSentinel/TripleClickOccured", "Touch");
   ros::shutdown();
+  delete nh_;
 }
 
-void Touch::spinTopics() {
-  this->checkBumpers();
-  this->checkChest();
-  this->checkHead();
-  this->checkHands();
+void Touch::init() {
+  try {
+    fMemoryProxy = AL::ALMemoryProxy(getParentBroker());
+    fMemoryProxy.subscribeToEvent("RightBumperPressed", "Touch",
+                                  "rightBumperPressed");
+    fMemoryProxy.subscribeToEvent("LeftBumperPressed", "Touch",
+                                  "leftBumperPressed");
+    fMemoryProxy.subscribeToEvent("FrontTactilTouched", "Touch",
+                                  "frontHeadPressed");
+    fMemoryProxy.subscribeToEvent("MiddleTactilTouched", "Touch",
+                                  "middleHeadPressed");
+    fMemoryProxy.subscribeToEvent("RearTactilTouched", "Touch",
+                                  "rearHeadPressed");
+    fMemoryProxy.subscribeToEvent("HandRightBackTouched", "Touch",
+                                  "handRightBackPressed");
+    fMemoryProxy.subscribeToEvent("HandRightLeftTouched", "Touch",
+                                  "handRightLeftPressed");
+    fMemoryProxy.subscribeToEvent("HandRightRightTouched", "Touch",
+                                  "handRightRightPressed");
+    fMemoryProxy.subscribeToEvent("HandLeftBackTouched", "Touch",
+                                  "handLeftBackPressed");
+    fMemoryProxy.subscribeToEvent("HandLeftLeftTouched", "Touch",
+                                  "handLeftLeftPressed");
+    fMemoryProxy.subscribeToEvent("HandLeftRightTouched", "Touch",
+                                  "handLeftRightPressed");
+    fMemoryProxy.subscribeToEvent("ALSentinel/SimpleClickOccured", "Touch",
+                                  "singleChestClick");
+    fMemoryProxy.subscribeToEvent("ALSentinel/DoubleClickOccured", "Touch",
+                                  "doubleChestClick");
+    fMemoryProxy.subscribeToEvent("ALSentinel/TripleClickOccured", "Touch",
+                                  "tripleChestClick");
+  } catch (const AL::ALError& e) {
+    DEBUG(e.what() << std::endl);
+  }
 }
 
-void Touch::checkBumpers() {
-  sensing::bumpers msg;
-  if (float(memProxy_->getData("RightBumperPressed")) > 0.5) {
-    msg.right = true;
+
+void Touch::rightBumperPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("RightBumperPressed");
+  if (fState  > 0.5f) {
+    bumperMsg_.right = true;
   } else {
-    msg.right = false;
+    bumperMsg_.right = false;
   }
-  if (float(memProxy_->getData("LeftBumperPressed")) > 0.5) {
-    msg.left = true;
-  } else {
-    msg.left = false;
-  }
-  bumpers_pub_.publish(msg);
+  bumpers_pub_.publish(bumperMsg_);
 }
 
-void Touch::checkChest() {
-  std_msgs::Bool msg;
-  if (float(memProxy_->getData("ChestButtonPressed")) > 0.5) {
-    msg.data = true;
+void Touch::leftBumperPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("LeftBumperPressed");
+  if (fState  > 0.5f) {
+    bumperMsg_.left = true;
   } else {
-    msg.data = false;
+    bumperMsg_.left = false;
   }
-  chest_pub_.publish(msg);
+  bumpers_pub_.publish(bumperMsg_);
 }
 
-void Touch::checkHead() {
-  sensing::head msg;
-  if (float(memProxy_->getData("FrontTactilTouched")) > 0.5) {
-    msg.front = true;
-  } else {
-    msg.front = false;
-  }
-  if (float(memProxy_->getData("MiddleTactilTouched")) > 0.5) {
-    msg.middle = true;
-  } else {
-    msg.middle = false;
-  }
-  if (float(memProxy_->getData("RearTactilTouched")) > 0.5) {
-    msg.rear = true;
-  } else {
-    msg.rear = false;
-  }
-  head_pub_.publish(msg);
+void Touch::singleChestClick() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  chestMsg_.data = "single";
+  chest_pub_.publish(chestMsg_);
 }
 
-void Touch::checkHands() {
-  this->checkRightHand();
-  this->checkLeftHand();
+void Touch::doubleChestClick() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  chestMsg_.data = "double";
+  chest_pub_.publish(chestMsg_);
 }
 
-void Touch::checkRightHand() {
-  sensing::hand msg;
-  if (float(memProxy_->getData("HandRightBackTouched")) > 0.5) {
-    msg.back = true;
-  } else {
-    msg.back = false;
-  }
-  if (float(memProxy_->getData("HandRightLeftTouched")) > 0.5) {
-    msg.left = true;
-  } else {
-    msg.left = false;
-  }
-  if (float(memProxy_->getData("HandRightRightTouched")) > 0.5) {
-    msg.right = true;
-  } else {
-    msg.right = false;
-  }
-  right_hand_pub_.publish(msg);
+void Touch::tripleChestClick() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  chestMsg_.data = "triple";
+  chest_pub_.publish(chestMsg_);
 }
 
-void Touch::checkLeftHand() {
-  sensing::hand msg;
-  if (float(memProxy_->getData("HandLeftBackTouched")) > 0.5) {
-    msg.back = true;
+void Touch::frontHeadPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("FrontTactilTouched");
+  if (fState  > 0.5f) {
+    headMsg_.front = true;
   } else {
-    msg.back = false;
+    headMsg_.front = false;
   }
-  if (float(memProxy_->getData("HandLeftLeftTouched")) > 0.5) {
-    msg.left = true;
-  } else {
-    msg.left = false;
-  }
-  if (float(memProxy_->getData("HandLeftRightTouched")) > 0.5) {
-    msg.right = true;
-  } else {
-    msg.right = false;
-  }
-  left_hand_pub_.publish(msg);
+  head_pub_.publish(headMsg_);
 }
+
+void Touch::middleHeadPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("MiddleTactilTouched");
+  if (fState  > 0.5f) {
+    headMsg_.middle = true;
+  } else {
+    headMsg_.middle = false;
+  }
+  head_pub_.publish(headMsg_);
+}
+
+void Touch::rearHeadPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("RearTactilTouched");
+  if (fState  > 0.5f) {
+    headMsg_.rear = true;
+  } else {
+    headMsg_.rear = false;
+  }
+  head_pub_.publish(headMsg_);
+}
+
+void Touch::handRightBackPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandRightBackTouched");
+  if (fState  > 0.5f) {
+    rightHandMsg_.back = true;
+  } else {
+    rightHandMsg_.back = false;
+  }
+  right_hand_pub_.publish(rightHandMsg_);
+}
+
+void Touch::handRightLeftPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandRightLeftTouched");
+  if (fState  > 0.5f) {
+    rightHandMsg_.left = true;
+  } else {
+    rightHandMsg_.left = false;
+  }
+  right_hand_pub_.publish(rightHandMsg_);
+}
+
+void Touch::handRightRightPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandRightRightTouched");
+  if (fState  > 0.5f) {
+    rightHandMsg_.right = true;
+  } else {
+    rightHandMsg_.right = false;
+  }
+  right_hand_pub_.publish(rightHandMsg_);
+}
+
+void Touch::handLeftBackPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandLeftBackTouched");
+  if (fState  > 0.5f) {
+    leftHandMsg_.back = true;
+  } else {
+    leftHandMsg_.back = false;
+  }
+  left_hand_pub_.publish(leftHandMsg_);
+}
+
+void Touch::handLeftLeftPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandLeftLeftTouched");
+  if (fState  > 0.5f) {
+    leftHandMsg_.left = true;
+  } else {
+    leftHandMsg_.left = false;
+  }
+  left_hand_pub_.publish(leftHandMsg_);
+}
+
+void Touch::handLeftRightPressed() {
+  AL::ALCriticalSection section(fCallbackMutex);
+  fState =  fMemoryProxy.getData("HandLeftRightTouched");
+  if (fState  > 0.5f) {
+    leftHandMsg_.right = true;
+  } else {
+    leftHandMsg_.right = false;
+  }
+  left_hand_pub_.publish(leftHandMsg_);
+}
+
