+++ edante/sensing/src/touch.h
@@ -1,8 +1,8 @@
 /*
 * @Copyright: Copyright[2015]<Alejandro Bordallo>
-* @Date:      2015-05-12
+* @Date:      2015-05-11
 * @Email:     alex.bordallo@ed.ac.uk
-* @Desc:      ROS wrapper for Nao's touch sensors
+* @Desc:      ROS wrapper for NaoQI Fall Manager API
 */
 
 #ifndef TOUCH_H
@@ -10,33 +10,65 @@
 
 #include <ros/ros.h>
 #include <std_msgs/Bool.h>
-
+#include <std_msgs/String.h>
 #include "sensing/bumpers.h"
-#include "sensing/head.h"
 #include "sensing/hand.h"
+#include "sensing/head.h"
 
+#include <alcommon/almodule.h>
+#include <alproxies/almotionproxy.h>
 #include <alproxies/almemoryproxy.h>
 
 #include "definitions.h"
 
+#include <boost/shared_ptr.hpp>
+#include <alcommon/almodule.h>
+#include <string>
+
+#include <alproxies/almemoryproxy.h>
+#include <alproxies/altexttospeechproxy.h>
+#include <althread/almutex.h>
+
+namespace AL {
+class ALBroker;
+}
+
 using namespace std;
 
-class Touch {
+class Touch : public AL::ALModule {
+
  public:
-  Touch(ros::NodeHandle* nh, AL::ALMemoryProxy* memProxy);
-  ~Touch();
 
-  // ROS publishers
-  void spinTopics();
+  Touch(boost::shared_ptr<AL::ALBroker> broker, const std::string& name);
+  virtual ~Touch();
 
-  void checkBumpers();
-  void checkChest();
-  void checkHead();
-  void checkHands();
-  void checkRightHand();
-  void checkLeftHand();
+  /** Overloading ALModule::init().
+  * This is called right after the module has been loaded
+  */
+  virtual void init();
+
+  void rightBumperPressed();
+  void leftBumperPressed();
+  void singleChestClick();
+  void doubleChestClick();
+  void tripleChestClick();
+  void frontHeadPressed();
+  void middleHeadPressed();
+  void rearHeadPressed();
+  void handRightBackPressed();
+  void handRightLeftPressed();
+  void handRightRightPressed();
+  void handLeftBackPressed();
+  void handLeftLeftPressed();
+  void handLeftRightPressed();
 
  private:
+  AL::ALMemoryProxy fMemoryProxy;
+
+  boost::shared_ptr<AL::ALMutex> fCallbackMutex;
+
+  float fState;
+
   // ROS
   ros::NodeHandle* nh_;
   ros::Publisher bumpers_pub_;
@@ -44,9 +76,14 @@ class Touch {
   ros::Publisher head_pub_;
   ros::Publisher right_hand_pub_;
   ros::Publisher left_hand_pub_;
+  sensing::bumpers bumperMsg_;
+  std_msgs::String chestMsg_;
+  sensing::head headMsg_;
+  sensing::hand rightHandMsg_;
+  sensing::hand leftHandMsg_;
 
   // NaoQI
   AL::ALMemoryProxy* memProxy_;
 };
 
-#endif  /* TOUCH_H */
+#endif /* TOUCH_H_ */
